* Replace uint16_t  with uint32_t  to avoid issues with large mana amount (65k+)

+++ src/creatures/creature.hpp
@@ -217,19 +217,19 @@ public:
 		return mana;
 	}
 
-	uint16_t getManaShield() const {
+	uint32_t getManaShield() const {
 		return manaShield;
 	}
 
-	void setManaShield(uint16_t value) {
+	void setManaShield(uint32_t value) {
 		manaShield = value;
 	}
 
-	uint16_t getMaxManaShield() const {
+	uint32_t getMaxManaShield() const {
 		return maxManaShield;
 	}
 
-	void setMaxManaShield(uint16_t value) {
+	void setMaxManaShield(uint32_t value) {
 		maxManaShield = value;
 	}
 
@@ -756,7 +756,7 @@ protected:
 	int32_t health = 1000;
 	int32_t healthMax = 1000;
 
-	uint16_t manaShield = 0;
+	uint32_t manaShield = 0;
 	uint16_t maxManaShield = 0;
 	int32_t varBuffs[BUFF_LAST + 1] = { 100, 100, 100 };
